fix(editor): text-fit runs in DOM renderer so preview matches generated image
ROOT CAUSE: CanvasContent.TextLayerEl used layer.font_size unconditionally
(line 148 of previous version). The fitFontSize binary-search algorithm only
ran in lib/image/compositing/layer-renderer.ts (server-only). With text_fit
enabled, the editor showed text at the static font_size while the generated
sharp image showed the binary-search-fitted size — they differed.

FIX:
  1. Extract fitFontSize, wrapLayerText, measureTextWidth, CHAR_RATIO from
     layer-renderer.ts into lib/image/text-fit-utils.ts — no server-only
     import, safe for both client (DOM renderer) and server (sharp renderer).
  2. layer-renderer.ts now imports from text-fit-utils.ts and re-exports for
     backward compat with existing test imports.
  3. CanvasContent.TextLayerEl: when text_fit.enabled, calls fitFontSize
     before building textStyle. Same algorithm, same inputs, same output as
     the sharp renderer.

Golden-image suite: 9/9 pass (server renderer unchanged). Unit tests: 273/273.
